+++ src/type/entity.ts
@@ -123,6 +123,8 @@ export type Dataset = BaseEntity &
     managerName?: string
     nbVariable?: number
     nextUpdateDate?: string
+    keyVariables?: Variable[]
+    businessKeyVariables?: Variable[]
     fkDatasetIds?: Set<string | number>
     fkReferencedByDatasetIds?: Set<string | number>
   }
@@ -143,6 +145,7 @@ export type Variable = BaseEntity &
     conceptId?: string | number
     originalName?: string
     key?: string | boolean
+    businessKey?: string | boolean
     nbDistinct?: number
     nbDuplicate?: number
     nbMissing?: number
@@ -333,6 +336,7 @@ export type MetaVariable = Omit<BaseEntity, 'id'> & {
   folderName?: string
   metaLocalisation?: string
   key?: string
+  businessKey?: string
 }
 
 export type MetaDataset = Omit<BaseEntity, 'id' | 'description'> & {